unifies newsletter and share button styles, moves them into a single component
