+++ src/App.css
@@ -79,7 +79,7 @@ tr {
 
 .Map path:hover {
   opacity: 0.5;
-  fill: rgba(0,0,0,0.25);
+  fill: rgba(0, 0, 0, 0.25);
   cursor: pointer;
 }
