+++ src/screens/FAQ/FAQ.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 import { TEAM } from './../../enums';
-import Newsletter from 'components/Newsletter/Newsletter';
+import ShareBlock from 'components/ShareBlock/ShareBlock';
 
 import { Wrapper, Content } from './FAQ.style';
 
@@ -498,7 +498,7 @@ const FAQ = ({ children }) => {
           next three weeks and will be distributed automatically, with no action
           required for most people.” .
         </Typography>
-        <Newsletter />
+        <ShareBlock />
       </Content>
     </Wrapper>
   );
